[feat] 룸 등록쪽 우선 커밋....
